requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
flask-mail>=0.10.0
gunicorn>=21.2.0
torch>=2.0.0
librosa>=0.10.0
openai-whisper>=20230314
yt-dlp>=2023.10.0
lyricsgenius>=3.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0
mutagen>=1.47.0
itsdangerous>=2.2.0
cryptography>=46.0.0
requests>=2.31.0
